Refactoring all

+++ MedView/voxel.cpp
@@ -4,17 +4,17 @@ Voxel::Voxel() {
 	Voxel(0.0, 0.0, 0.0, 0);
 }
 
-Voxel::Voxel(double x, double y, double z, int intensity = 0) {
-	this->setCoord(Eigen::Vector4d(x, y, z, 1.0));
+Voxel::Voxel(float x, float y, float z, int intensity = 0) {
+	this->setCoord(Eigen::Vector4f(x, y, z, 1.0));
 
 	this->m_intensity = intensity;
 }
 
-void Voxel::setCoord(Eigen::Vector4d pos) {
+void Voxel::setCoord(Eigen::Vector4f pos) {
 	this->m_coord = pos;
 }
 
-Eigen::Vector4d Voxel::getCoord() {
+Eigen::Vector4f Voxel::getCoord() {
 	return m_coord;
 }
